Update contribution guidelines
Update smaller things in the contribution guidelines to be more explicit. On top, I added a new changelog file. In the future, we want the changelog to be the source of truth for version numbers. To do that, we need a format that can easily be parsed. "Keep a changelog" is a quasi-standard. The old changelog does not really fit into that format though, hence I created a new file but kept the old one around for reference.